aclia_data/requirements.txt
fuzzywuzzy==0.18.0
lime==0.2.0.1
nltk==3.4.5
numpy==1.18.1
pandas==1.0.1
pdf2image==1.12.1
Pillow==7.0.0
psutil==5.6.7
py7zr==0.6
PyPDF2==1.26.0
pytesseract==0.3.3
seaborn==0.10.0
spacy==2.2.4
sqlalchemy==1.3.13
tika==1.23.1
xgboost==1.1.0